version 1.3.0 - A* as a service
