+++ AStar/osgi.bnd
@@ -1,0 +1,3 @@
+Bundle-SymbolicName: AStar
+Bundle-ActivationPolicy: lazy
+Service-Component: META-INF/searcher.xml
