+++ AStar/src/main/java/dk/sdu/mmmi/astar/Node.java
@@ -1,4 +1,4 @@
-package dk.sdu.mmmi.enemy;
+package dk.sdu.mmmi.astar;
 
 import dk.sdu.mmmi.common.data.entityparts.PositionPart;
 import java.util.ArrayList;
