+++ Common/src/main/java/dk/sdu/mmmi/common/services/ICollisionChecker.java
@@ -1,6 +1,5 @@
 package dk.sdu.mmmi.common.services;
 
 public interface ICollisionChecker {
-
     boolean isPositionFree(float x, float y);
 }
