+++ Enemy/src/main/resources/META-INF/entityprocessor.xml
@@ -5,7 +5,7 @@
         <provide interface="dk.sdu.mmmi.common.services.IEntityProcessingService"/>
     </service>
     
-    <reference bind="addCollisionChecker" cardinality="0..n"
-               interface="dk.sdu.mmmi.common.services.ICollisionChecker"
-               name="ICollisionChecker" policy="dynamic" unbind="removeCollisionChecker"/>
+    <reference bind="addSearcher" cardinality="0..n"
+               interface="dk.sdu.mmmi.common.services.ISearcher"
+               name="ISearcher" policy="dynamic" unbind="removeSearcher"/>
 </scr:component>
