+++ pom.xml
@@ -24,6 +24,7 @@
         <module>LibGDX</module> 
         <module>Enemy</module> 
         <module>Player</module> 
+        <module>AStar</module> 
     </modules> 
     <build> 
         <plugins> 
@@ -34,11 +35,10 @@
                 <configuration> <!-- | some example Pax-Runner settings --> 
                     <provision> 
                         <param>--platform=felix</param> 
-                        <param>--vmOptions=-agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=5005</param> 
                     </provision> 
                 </configuration> 
                 <executions> <!-- | uncomment to auto-generate IDE files <execution> <id>ide-support</id> <goals> <goal>eclipse</goal> </goals> </execution> --> </executions> 
             </plugin> 
         </plugins> 
     </build> 
-</project>
+</project>
