+++ runner/deploy-pom.xml
@@ -87,6 +87,13 @@
       <type>bundle</type>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>dk.sdu.mmmi.build</groupId>
+      <artifactId>AStar</artifactId>
+      <version>1.0-SNAPSHOT</version>
+      <type>bundle</type>
+      <scope>provided</scope>
+    </dependency>
   </dependencies>
   <repositories>
     <repository>
