+++ runner/felix/config.ini
@@ -32,7 +32,8 @@ felix.auto.start.5=\
 "file:bundles/dk.sdu.mmmi.build.LibGDX_1.0.0.SNAPSHOT.jar" \
 "file:bundles/Core_1.0.0.SNAPSHOT.jar" \
 "file:bundles/Enemy_1.0.0.SNAPSHOT.jar" \
-"file:bundles/Player_1.0.0.SNAPSHOT.jar"
+"file:bundles/Player_1.0.0.SNAPSHOT.jar" \
+"file:bundles/AStar_1.0.0.SNAPSHOT.jar"
 felix.auto.install.5=\
 "file:bundles/com.diffplug.osgi.extension.sun.misc_0.0.0.jar" \
 "file:bundles/com.diffplug.osgi.extension.sun.reflect_0.0.0.jar"
